Added stuff like sign out button but layout still janky in map but button for search is black
